backend/requirements.txt
# Core Framework
fastapi==0.115.5
uvicorn[standard]==0.32.1
python-multipart==0.0.18

# Database
sqlalchemy==2.0.36
psycopg2-binary==2.9.10
alembic==1.14.0

# Data Validation & Serialization
pydantic==2.10.3
pydantic-settings==2.6.1
email-validator==2.2.0

# Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.1

# Utilities
python-dateutil==2.9.0.post0
pytz==2024.2

# Development & Testing
pytest==8.3.4
pytest-asyncio==0.24.0
httpx==0.28.1

# CORS
fastapi-cors==0.0.6